Update go-ipfs
Corrected the missing &gt; when appending the 2nd line to root's .profile. 
Also the use of runuser to execute the ipfs init as the ipfs account will fail if it doesn't run the ipfs account's .profile script. The manpage for runuser is not clear about that, but env should be correct and -l simulates a login as the target user so it should work.