Create go-ipfs-amd64
This is essentially the same as go-ipfs, but for Debian servers. They should be combined at some point as the differences aren't that great.

Biggest issue with using either of these go-ipfs scripts is that the golang version changes frequently as does the ipfs-update, and I've seen errors when those things aren't compatible. This requires these scripts be edited to make sure the golang version is compatible with the ipfs & ipfs-update versions available.